f T34551 refactor DeleteProcedureCommand
- add ProcedureDeleter
